+++ app/resources/views/pages/dominion/calculations/military.blade.php
@@ -735,7 +735,7 @@
 
                             <div class="form-group row">
                                 @php
-                                    $racialSpell = $spellHelper->getSpellsWithPerk(['offense', 'offense_from_barren_land', 'offense_from_spell', 'offense_from_pairing_demon'], $race)->first();
+                                    $racialSpell = $spellHelper->getSpellsWithPerk(['offense', 'offense_from_barren_land', 'offense_from_spell', 'offense_unit1'], $race)->first();
                                 @endphp
                                 <div class="col-xs-3 text-right">
                                     @if ($racialSpell)