+++ app/resources/views/pages/dominion/invade.blade.php
@@ -143,7 +143,7 @@
                                                        data-amount="{{ $selectedDominion->{"military_unit{$unitSlot}"} }}"
                                                        data-op="{{ $unit->power_offense }}"
                                                        data-dp="{{ $unit->power_defense }}"
-                                                       data-need-boat="{{ (int)$unit->need_boat }}"
+                                                       data-need-boat="{{ $militaryCalculator->getUnitNeedBoats($dominion, $unit) }}"
                                                        {{ $selectedDominion->isLocked() ? 'disabled' : null }}>
                                             </td>
                                             <td class="text-center" id="unit{{ $unitSlot }}_stats">