+++ src/Services/Dominion/Actions/RaidActionService.php
@@ -378,7 +378,7 @@ class RaidActionService
                 continue;
             }
 
-            if ($unit->need_boat) {
+            if ($this->militaryCalculator->getUnitNeedBoats($dominion, $unit)) {
                 $hours = $this->invasionService->getUnitReturnHoursForSlot($dominion, $unit->slot);
 
                 if (!isset($unitsThatNeedsBoatsByReturnHours[$hours])) {